requirements.txt
pyyaml==5.4.1
bootstrap-flask==1.8.0
flask==2.0.1
gunicorn==20.1.0
flask-mongoengine==1.0.0

